+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 #### 0.0.3-alpha - Friday, September 4, 2020 (WIP)
 - added Surprisal Analysis cp Heatmap Plot
+- added Surprisal Analysis analytic plots: 
+  - Constraint importance
+  - Gradual data recovery
 
 #### 0.0.2-alpha - Wednesday, September 2, 2020
 - added Plot functions for Surprisal Analysis results.